+++ apps/commission/admin.py
@@ -5,6 +5,7 @@ from apps.commission.models import (
     Commission,
     CommissionFile,
     CommissionItem,
+    CommissionNote,
 )
 
 
@@ -13,3 +14,4 @@ admin.site.register(Vehicle)
 admin.site.register(Commission)
 admin.site.register(CommissionItem)
 admin.site.register(CommissionFile)
+admin.site.register(CommissionNote)