+++ apps/commission/graphql/mutations.py
@@ -1,5 +1,5 @@
 from KlimaKar.graphql.resolvers import BaseModelFormResolver
-from apps.commission.models import Commission
+from apps.commission.models import Commission, CommissionNote
 from apps.commission.forms import (
     CommissionModelForm,
     CommissionItemModelForm,
@@ -11,8 +11,8 @@ from apps.commission.graphql.types import mutation
 
 
 @mutation.field("updateStatus")
-def resolve_update_status(_, info, id, status):
-    c = Commission.objects.get(id=id)
+def resolve_update_status(_, info, pk, status):
+    c = Commission.objects.get(id=pk)
     c.status = status.value
     c.save()
     return c
@@ -45,3 +45,25 @@ class AddVehicleResolver(BaseModelFormResolver):
 @mutation.field("addVehicle")
 def resolve_add_vehicle(_, info, data):
     return AddVehicleResolver(data).process()
+
+
+@mutation.field("addCommissionNote")
+def resolve_add_commission_note(_, info, commission, contents):
+    try:
+        commission_obj = Commission.objects.get(pk=commission)
+    except Commission.DoesNotExist:
+        return None
+    note = CommissionNote.objects.create(commission=commission_obj, contents=contents)
+    return note
+
+
+@mutation.field("updateCommissionNote")
+def resolve_update_commission_note(_, info, pk, contents, isActive):
+    try:
+        note = CommissionNote.objects.get(pk=pk)
+    except CommissionNote.DoesNotExist:
+        return None
+    note.contents = contents
+    note.is_active = isActive
+    note.save()
+    return note